requirements.txt
flask
flask-swagger-ui
pyyaml